+++ lib/src/intTest/java/blackbox/reader/IndexedCsvReaderTest.java
@@ -75,7 +75,7 @@ class IndexedCsvReaderTest {
 
         assertThat(singlePageBuilder().ofCsvRecord(file)).asString()
             .isEqualTo("IndexedCsvReader[file=%s, charset=UTF-8, fieldSeparator=,, "
-                    + "quoteCharacter=\", commentStrategy=NONE, commentCharacter=#, acceptCharsAfterQuotes=true, "
+                    + "quoteCharacter=\", commentStrategy=NONE, commentCharacter=#, acceptCharsAfterQuotes=false, "
                     + "pageSize=1, "
                     + "index=CsvIndex[bomHeaderLength=0, fileSize=3, fieldSeparator=44, quoteCharacter=34, "
                     + "commentStrategy=NONE, commentCharacter=35, recordCount=1, pageCount=1]]",
@@ -146,7 +146,8 @@ class IndexedCsvReaderTest {
 
     @Test
     void acceptCharsAfterQuotes() throws IOException {
-        try (var csv = singlePageBuilder().ofCsvRecord(prepareTestFile("foo,\"bar\"baz"), StandardCharsets.UTF_8)) {
+        final var bldr = singlePageBuilder().acceptCharsAfterQuotes(true);
+        try (var csv = bldr.ofCsvRecord(prepareTestFile("foo,\"bar\"baz"), StandardCharsets.UTF_8)) {
             CsvRecordAssert.assertThat(csv.readPage(0).getFirst()).fields()
                 .containsExactly("foo", "barbaz");
         }
@@ -154,8 +155,7 @@ class IndexedCsvReaderTest {
 
     @Test
     void acceptCharsAfterQuotesNot() throws IOException {
-        final var bldr = singlePageBuilder().acceptCharsAfterQuotes(false);
-        try (var csv = bldr.ofCsvRecord(prepareTestFile("foo,\"bar\"baz"), StandardCharsets.UTF_8)) {
+        try (var csv = singlePageBuilder().ofCsvRecord(prepareTestFile("foo,\"bar\"baz"), StandardCharsets.UTF_8)) {
             assertThatThrownBy(() -> csv.readPage(0))
                 .isInstanceOf(CsvParseException.class);
         }