+++ lib/src/main/java/de/siegmar/fastcsv/reader/CsvParser.java
@@ -3,7 +3,7 @@ package de.siegmar.fastcsv.reader;
 import java.io.Closeable;
 import java.io.IOException;
 
-interface CsvParser extends Closeable {
+sealed interface CsvParser extends Closeable permits LooseCsvParser, StrictCsvParser {
 
     /// {@return `true` if data was parsed, `false` if the end of the stream was reached.}
     boolean parse() throws IOException;
@@ -13,7 +13,8 @@ interface CsvParser extends Closeable {
 
     /// Skips a line in the stream after skipping the specified number of characters.
     ///
-    /// @return `true` if the line was skipped, `false` if the end of the stream was reached before any characters could be skipped.
+    /// @return `true` if the line was skipped,
+    ///     `false` if the end of the stream was reached before any characters could be skipped.
     boolean skipLine(int numCharsToSkip) throws IOException;
 
     /// {@return the starting line number of the last parsed record.}